Add category and sort out
