+++ app/src/main/java/com/kantai_wiki/kcw_mobile/activity/MainActivity.java
@@ -1,4 +1,4 @@
-package com.kantai_wiki.kcw_mobile;
+package com.kantai_wiki.kcw_mobile.activity;
 
 import android.os.Bundle;
 import android.support.design.widget.TabLayout;
@@ -7,6 +7,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import com.kantai_wiki.kcw_mobile.R;
 
 public class MainActivity extends AppCompatActivity {
