+++ app/src/main/java/com/kantai_wiki/kcw_mobile/activity/ShipAdapter.java
@@ -1,4 +1,4 @@
-package com.kantai_wiki.kcw_mobile;
+package com.kantai_wiki.kcw_mobile.activity;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
@@ -6,6 +6,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.kantai_wiki.kcw_mobile.R;
 
 import java.util.ArrayList;
 import java.util.List;
