+++ app/src/main/java/com/kantai_wiki/kcw_mobile/activity/TabFragment_attack.java
@@ -1,4 +1,4 @@
-package com.kantai_wiki.kcw_mobile;
+package com.kantai_wiki.kcw_mobile.activity;
 
 /**
  * Created by airfr on 2015/8/25.
@@ -11,6 +11,8 @@ import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.Toast;
+
+import com.kantai_wiki.kcw_mobile.R;
 
 import java.util.ArrayList;
