+++ app/src/main/java/com/kantai_wiki/kcw_mobile/activity/TabFragment_equipment.java
@@ -1,4 +1,4 @@
-package com.kantai_wiki.kcw_mobile;
+package com.kantai_wiki.kcw_mobile.activity;
 
 /**
  * Created by airfr on 2015/8/25.
@@ -9,13 +9,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.kantai_wiki.kcw_mobile.R;
+
 
 /**
  * Created by airfr on 2015/8/24.
  */
-public class TabFragment_expedition extends Fragment {
+public class TabFragment_equipment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle saveInstanceState){
-        return inflater.inflate(R.layout.tab_fragment_expedition,container,false);
+        return inflater.inflate(R.layout.tab_fragment_equipment,container,false);
     }
 }
